fix #5: Move workerskv to mod directory
* Key and WorkersKVNamespace derive Clone, PartialEq, Eq
