+++ src/workerskv/create_namespace.rs
@@ -1,0 +1,26 @@
+use super::WorkersKVNamespace;
+
+use crate::endpoint::{Endpoint, Method};
+
+/// Create a Namespace
+/// Creates a namespace under the given title.
+/// A 400 is returned if the account already owns a namespace with this title.
+/// A namespace must be explicitly deleted to be replaced.
+/// https://api.cloudflare.com/#workers-kv-namespace-create-a-namespace
+pub struct CreateNamespace<'a> {
+    pub account_identifier: &'a str,
+}
+
+impl<'a> Endpoint<WorkersKVNamespace, CreateNamespaceParams> for CreateNamespace<'a> {
+    fn method(&self) -> Method {
+        Method::Post
+    }
+    fn path(&self) -> String {
+        format!("accounts/{}/storage/kv/namespaces", self.account_identifier)
+    }
+}
+
+#[derive(Serialize, Clone, Debug)]
+pub struct CreateNamespaceParams {
+    pub title: String,
+}
